Add new billing presenter and move existing to it
