+++ app/presenters/bill-runs/index-bill-runs.presenter.js
@@ -5,7 +5,8 @@
  * @module IndexBillRunsPresenter
  */
 
-const { formatBillRunType, formatLongDate, formatMoney, titleCase } = require('../base.presenter.js')
+const { formatLongDate, formatMoney, titleCase } = require('../base.presenter.js')
+const { formatBillRunType } = require('../billing.presenter.js')
 
 /**
  * Formats the summary data for each bill run for use in the /bill-runs page
